+++ test/faker/tolkien_test.exs
@@ -1,0 +1,6 @@
+defmodule Faker.TolkienTest do
+  use ExUnit.Case, async: true
+
+  doctest Faker.Tolkien
+  doctest Faker.Tolkien.En
+end